Remove free chlorine; Update hints

+++ app/src/main/java/com/bach/gardenstate/features/sensors/model/SoilMoistureSensorData.kt
@@ -8,5 +8,5 @@ data class SoilMoistureSensorData(
     val last_seen: String,
     val linkquality: Int,
     val soil_moisture: Int,
-    val temperature: Int
+    val temperature: Float
 )
